Split views in differents files
